+++ EWSContacts/Module/internal/scripts/preimport.ps1
@@ -1,1 +1,0 @@
-﻿# Add all things you want to run before importing the main code