+++ EWSContacts/Module/tests/pester.ps1
@@ -1,1 +1,0 @@
-﻿# Insert tests here